Removing basemodel from patch types. Moving transaction types back to extensions.transaction. Adding literals for content_type. Removing sys check for TypedDict import.
